fix: switch colpali_compat backward parity to fp16 to match kernel dtype
bf16 still produced ~0.39 max-abs relative gradient diff because its 7
mantissa bits leave too many near-tied dot-product values where the
kernel's fp32-domain argmax disagrees with einsum's bf16-domain argmax.
fp16 has 10 mantissa bits, mirrors the dtype used by the passing
test_patched_loss_forwards_match_original_cuda, and tightens the tie
margin enough to bring gradient parity back to 5e-3.

+++ tests/test_colpali_compat.py
@@ -383,13 +383,16 @@ def test_patched_colbert_loss_backward_matches_original(fake_colpali):
     ColbertLoss = losses.ColbertLoss  # noqa: N806
 
     torch.manual_seed(0)
-    # bf16 matches the kernel's internal compute dtype (`pick_compute_dtype`
-    # forces fp16/bf16 even when callers pass fp32). With fp32 inputs the
-    # autograd reference runs in fp32 while the kernel quantizes to bf16
-    # — on L2-normalized tokens that's enough to flip the inner argmax on
-    # near-tied max candidates and blow up the gradient diff.
-    Q0 = _l2(torch.randn(8, 32, 128, device="cuda", dtype=torch.bfloat16))
-    D0 = _l2(torch.randn(8, 96, 128, device="cuda", dtype=torch.bfloat16))
+    # fp16 matches the kernel's internal compute dtype (`pick_compute_dtype`
+    # picks fp16 unless either input is bf16). With fp32 inputs the autograd
+    # reference runs in fp32 while the kernel quantizes to fp16 — on
+    # L2-normalized tokens that's enough to flip the inner argmax on
+    # near-tied max candidates and blow up the gradient diff. fp16 (over
+    # bf16) gives tighter argmax agreement here because it has 10 mantissa
+    # bits vs bf16's 7, so fewer ties at the bottom of the dot-product
+    # distribution. Mirrors `test_patched_loss_forwards_match_original_cuda`.
+    Q0 = _l2(torch.randn(8, 32, 128, device="cuda", dtype=torch.float16))
+    D0 = _l2(torch.randn(8, 96, 128, device="cuda", dtype=torch.float16))
 
     head = ColbertLoss(normalize_scores=False).to("cuda")
